+++ smithinnovates/src/frontend/SearchedResults.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function SearchedResults() {
+  return (
+    <div>
+      <h1>Results</h1>
+    </div>
+  )
+}
+
+export default SearchedResults
